Error handling: Create a base for error propagation (#13)
Error messages should be understandable by an end user, and instruct
them on what to try next or what failed.

Where the database or program entered an invalid state at no fault of
the user, no error is necessary, use `expect()` instead.

Errors should be propagated to a place where a way of printing it is
feasible. Use ui::display_error() wherever possible.